WINDUP-1684 Tech stats boxes report WIP, sorting out to boxes works, refactoring 1
