+++ reporting/impl/src/main/java/org/jboss/windup/reporting/rules/generation/techreport/GetTechnologiesIdentifiedForSubSectorAndRowMethod.java
@@ -1,6 +1,5 @@
 package org.jboss.windup.reporting.rules.generation.techreport;
 
-import com.tinkerpop.blueprints.Graph;
 import freemarker.ext.beans.StringModel;
 import freemarker.template.TemplateModelException;
 import java.util.*;
@@ -8,7 +7,6 @@ import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 import org.jboss.windup.config.GraphRewrite;
-import org.jboss.windup.config.tags.Tag;
 import org.jboss.windup.graph.GraphContext;
 import org.jboss.windup.graph.model.ProjectModel;
 import org.jboss.windup.reporting.freemarker.WindupFreeMarkerMethod;
@@ -16,7 +14,6 @@ import org.jboss.windup.reporting.model.TagModel;
 import org.jboss.windup.reporting.model.TechnologyUsageStatisticsModel;
 import org.jboss.windup.reporting.service.TagGraphService;
 import org.jboss.windup.util.ExecutionStatistics;
-import org.jboss.windup.util.exception.WindupException;
 
 /**
  * Gets the list of TechnologyUsageStatisticsModel-s which should be displayed in the box given by the report "coordinates" tags (subsector/box, row).
@@ -85,7 +82,7 @@ public class GetTechnologiesIdentifiedForSubSectorAndRowMethod implements Windup
                 projectModel = (ProjectModel) projectArg.getWrappedObject();
         }
 
-        Map<String, TechUsageStatSum> techStats = getTechStats(boxTag, rowTag, projectModel);
+        Map<String, TechReportService.TechUsageStatSum> techStats = getTechStats(boxTag, rowTag, projectModel);
 
         ExecutionStatistics.get().end(NAME);
         return techStats;
@@ -95,68 +92,41 @@ public class GetTechnologiesIdentifiedForSubSectorAndRowMethod implements Windup
 
     /**
      * This scans all {@link TechnologyUsageStatisticsModel}s and filters those belonging under given box/column and row, and project.
+     * @deprecated Remove.
      */
-    private Map<String, TechUsageStatSum> getTechStats(TagModel boxTag, TagModel rowTag, ProjectModel project)
+    private Map<String, TechReportService.TechUsageStatSum> getTechStats(TagModel boxTag, TagModel rowTag, ProjectModel project)
     {
         LOG.info(String.format("#### boxTag %s, rowTag %s, project %s", boxTag, rowTag, project));
 
         final TagGraphService tagService = new TagGraphService(graphContext);
 
-        Map<String, TechUsageStatSum> sums = new HashMap<>();
+        Map<String, TechReportService.TechUsageStatSum> sums = new HashMap<>();
 
         final Iterable<TechnologyUsageStatisticsModel> statModels = graphContext.service(TechnologyUsageStatisticsModel.class).findAll();
         final Set<TechnologyUsageStatisticsModel> forGivenBoxAndRow = StreamSupport.stream(statModels.spliterator(), false)
                 // Only the given project.
                 .filter(stat -> project == null || stat.getProjectModel() != null && stat.getProjectModel().asVertex().getId() == project.asVertex().getId()) /// Can models use equals?
-                .peek(stat -> LOG.info(String.format("    Checking '%s', so far %sx, tags: %s", stat.getName(), sums.getOrDefault(stat.getName(), new TechUsageStatSum("")).getOccurrenceCount(), stat.getTags())))
+                .peek(stat -> LOG.info(String.format("    Checking '%s', so far %sx, tags: %s", stat.getName(), sums.getOrDefault(stat.getName(), new TechReportService.TechUsageStatSum("")).getOccurrenceCount(), stat.getTags())))
                 // Only those under both row and box tags.
                 .filter(stat -> {
-                    final Set<String>[] normalAndSilly = splitSillyTagNames(graphContext, stat.getTags());
+                    final Set<String>[] normalAndSilly = TechReportService.splitSillyTagNames(graphContext, stat.getTags());
 
                     // Normal: any of the tags must fit into the given row and box/column,
                     if (anyTagsUnderAllTags(normalAndSilly[0], Arrays.asList(boxTag, rowTag)))
                         return true;
 
                     // Silly: there must be two silly labels representing row or box/column, or one label fitting the box name.
-                    TechReportPlacement placement = processSillyLabels(graphContext, normalAndSilly[1]);
-                    //return placement.box != null || (placement.sector != null && placement.box != null);
+                    TechReportService.TechReportPlacement placement = TechReportService.processSillyLabels(graphContext, normalAndSilly[1]);
                     return placementBelongToThisBoxAndRow(placement, boxTag, rowTag);
                 })
                 .peek(stat -> {
-                    LOG.info(String.format("    Summing '%s', so far %sx, tags: %s", stat.getName(), sums.getOrDefault(stat.getName(), new TechUsageStatSum("")).getOccurrenceCount(), stat.getTags()) );
-                    //sums.merge(stat.getName(), new TechUsageStatSum(stat), (sumFromMap, sumNext) -> sumFromMap.add(sumNext))///
+                    LOG.info(String.format("    Summing '%s', so far %sx, tags: %s", stat.getName(), sums.getOrDefault(stat.getName(), new TechReportService.TechUsageStatSum("")).getOccurrenceCount(), stat.getTags()) );
                     sums.put(stat.getName(),
-                            sums.getOrDefault(stat.getName(), new TechUsageStatSum(stat.getName())).add(stat) );
+                            sums.getOrDefault(stat.getName(), new TechReportService.TechUsageStatSum(stat.getName())).add(stat) );
                 })
                 .collect(Collectors.toSet());
 
-        //return forGivenBoxAndRow;
         return sums;
-    }
-
-    /**
-     * Translates the silly tags (labels) to their normalized real tag counterparts.
-     * Returns a 2-item array; index 0 has the normal names, index 1 the silly names.
-     *
-     * Due to bad design, the rules contain column and row titles for the graph, rather than technology tags.
-     * To make them fit into the tag system, this translation is needed. See also the "silly:..." tags in the report hierarchy definition.
-     */
-    static Set<String>[] splitSillyTagNames(GraphContext graphContext, Set<String> potentialSillyTags)
-    {
-        final TagGraphService tagService = new TagGraphService(graphContext);
-
-        Set<String> normalNames = new HashSet<>();
-        Set<String> sillyNames = new HashSet<>();
-
-        potentialSillyTags.stream().forEach(name -> {
-            final TagModel sillyTag = tagService.getTagByName("silly:" + Tag.normalizeName(name));
-            if (null != sillyTag)
-                sillyNames.add(sillyTag.getName());
-            else
-                // Some may be undefined. This will happen when someone attempts to add a new sector, row or column/box.
-                normalNames.add(sillyTag.getName());
-        });
-        return new Set[]{normalNames, sillyNames};
     }
 
     /**
@@ -188,169 +158,9 @@ public class GetTechnologiesIdentifiedForSubSectorAndRowMethod implements Windup
     /**
      * Returns whether out of three tags, one is under sectorTag and one under rowTag. The remaining one is supposedly the a box label.
      */
-    boolean placementBelongToThisBoxAndRow(TechReportPlacement placement, TagModel boxTag, TagModel rowTag)
+    boolean placementBelongToThisBoxAndRow(TechReportService.TechReportPlacement placement, TagModel boxTag, TagModel rowTag)
     {
         return tagService.isTagUnderTagOrSame(placement.box, boxTag) && tagService.isTagUnderTagOrSame(placement.row, rowTag);
     }
 
-
-    /**
-     * From three tagNames, if one is under sectorTag and one under rowTag, returns the remaining one, which is supposedly the a box label.
-     * Otherwise, returns null.
-     */
-    static TechReportPlacement processSillyLabels(GraphContext grCtx, Set<String> tagNames)
-    {
-        TagGraphService tagService = new TagGraphService(grCtx);
-
-        if (tagNames.size() < 3)
-            throw new WindupException("There should always be exactly 3 silly labels - row, sector, column/box. It was: " + tagNames);
-        if (tagNames.size() > 3)
-            LOG.severe("There should always be exactly 3 silly labels - row, sector, column/box. It was: " + tagNames);
-
-        TechReportPlacement placement = new TechReportPlacement();
-
-        final TagModel sillySectorsTag = tagService.getTagByName("techReport:sillySectors");
-        final TagModel sillyBoxesTag = tagService.getTagByName("techReport:sillyBoxes");
-        final TagModel sillyRowsTag = tagService.getTagByName("techReport:sillyRows");
-
-
-        Set<String> tagNames2 = new HashSet(tagNames);
-        for (Iterator<String> tagNamesIt = tagNames2.iterator(); tagNamesIt.hasNext(); )
-        {
-            String name = tagNamesIt.next();
-            final TagModel tag = tagService.getTagByName(name);
-            if (null == tag)
-                continue;
-
-            if (tagService.isTagUnderTagOrSame(tag, sillySectorsTag))
-            {
-                placement.sector = tag;
-                tagNamesIt.remove();
-            }
-            else if (tagService.isTagUnderTagOrSame(tag, sillyBoxesTag))
-            {
-                placement.box = tag;
-                tagNamesIt.remove();
-            }
-            else if (tagService.isTagUnderTagOrSame(tag, sillyRowsTag))
-            {
-                placement.row = tag;
-                tagNamesIt.remove();
-            }
-        }
-        placement.unknown = tagNames2;
-
-        LOG.info(String.format("\t\tLabels %s identified as: sector: %s, box: %s, row: %s", tagNames, placement.sector, placement.box, placement.row));
-        if (placement.box == null || placement.row == null)
-        {
-            LOG.severe(String.format("There should always be exactly 3 silly labels - row, sector, column/box. Found: %s, of which box: %s, row: %s", tagNames, placement.box, placement.row));
-        }
-        return placement;
-    }
-
-    /**
-     * This relies on the tag structure in the XML when the silly mapping tags have exactly one parent outside the silly group, which is the tag they are mapped to.
-     */
-    static TechReportPlacement normalizeSillyPlacement(GraphContext grCtx, TechReportPlacement sillyPlacement)
-    {
-        TagGraphService tagService = new TagGraphService(grCtx);
-
-        final TechReportPlacement normalPlacement = new TechReportPlacement();
-        normalPlacement.sector = getNonSillyParent(tagService, sillyPlacement.sector);
-        normalPlacement.box = getNonSillyParent(tagService, sillyPlacement.box);
-        normalPlacement.row = getNonSillyParent(tagService, sillyPlacement.row);
-        return normalPlacement;
-    }
-
-    private static TagModel getNonSillyParent(TagGraphService tagService, TagModel tag)
-    {
-        final TagModel sillyRoot = tagService.getTagByName("techReport:mappingOfSillyTagNames");
-
-        final Iterator<TagModel> parents = tag.getDesignatedByTags().iterator();
-        if (!parents.hasNext())
-            throw new WindupException("Tag is not designated by any tags: " + tag);
-
-        TagModel nonSillyParent = null;
-        do {
-            TagModel parentTag = parents.next();
-            if (tagService.isTagUnderTagOrSame(parentTag, sillyRoot))
-                continue;
-            if (nonSillyParent != null)
-                throw new WindupException(String.format("Tag %s has more than one non-silly parent: %s, %s", nonSillyParent, parentTag));
-            nonSillyParent = parentTag;
-        }
-        while (parents.hasNext());
-
-        return nonSillyParent;
-    }
-
-
-    public static class TechReportPlacement {
-        public TagModel sector;
-        public TagModel box;
-        public TagModel row;
-        public Set<String> unknown;
-
-        @Override
-        public String toString()
-        {
-            return "TechReportPlacement{sector=" + sector + ", box=" + box + ", row=" + row + ", unknown=" + unknown + '}';
-        }
-    }
-
-
-
-    /**
-     * Keeps the aggregated data from multiple {@link TechnologyUsageStatisticsModel}s.
-     */
-    public static class TechUsageStatSum {
-        String name;
-        int count = 0;
-        Set<String> tags = new HashSet<>();
-
-        public TechUsageStatSum(String name)
-        {
-            this.name = name;
-        }
-
-        public TechUsageStatSum(TechnologyUsageStatisticsModel stat)
-        {
-            this.name = stat.getName();
-            this.count = stat.getOccurrenceCount();
-            this.tags.addAll(stat.getTags());
-        }
-
-        public TechUsageStatSum add(TechnologyUsageStatisticsModel stat)
-        {
-            if (!this.name.equals(stat.getName()))
-                throw new IllegalArgumentException("Can't add up stats, " + this.name + " != " + stat.getName());
-            this.count += stat.getOccurrenceCount();
-            this.tags.addAll(stat.getTags());
-            return this;
-        }
-
-        public TechUsageStatSum add(TechUsageStatSum stat)
-        {
-            if (!this.name.equals(stat.getName()))
-                throw new IllegalArgumentException("Can't add up stats, " + this.name + " != " + stat.getName());
-            this.count += stat.getOccurrenceCount();
-            this.tags.addAll(stat.getTags());
-            return this;
-        }
-
-        public String getName()
-        {
-            return name;
-        }
-
-        public int getOccurrenceCount()
-        {
-            return count;
-        }
-
-        public Set<String> getTags()
-        {
-            return tags;
-        }
-    }
 }
