+++ reporting/impl/src/main/java/org/jboss/windup/reporting/rules/generation/techreport/SortTechUsageStatsMethod.java
@@ -6,19 +6,13 @@ import freemarker.template.SimpleScalar;
 import freemarker.template.TemplateModelException;
 import java.util.*;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 import org.jboss.windup.config.GraphRewrite;
-import org.jboss.windup.config.tags.Tag;
 import org.jboss.windup.graph.GraphContext;
 import org.jboss.windup.graph.model.ProjectModel;
 import org.jboss.windup.reporting.freemarker.WindupFreeMarkerMethod;
-import org.jboss.windup.reporting.model.TagModel;
 import org.jboss.windup.reporting.model.TechnologyUsageStatisticsModel;
-import static org.jboss.windup.reporting.rules.generation.techreport.GetTechnologiesIdentifiedForSubSectorAndRowMethod.*;
 import org.jboss.windup.reporting.service.TagGraphService;
 import org.jboss.windup.util.ExecutionStatistics;
-import org.jboss.windup.util.exception.WindupException;
 
 /**
  * Gets the list of TechnologyUsageStatisticsModel-s which should be displayed in the box given by the report "coordinates" tags (subsector/box, row).
@@ -84,7 +78,7 @@ public class SortTechUsageStatsMethod implements WindupFreeMarkerMethod
                 projectModel = (ProjectModel) projectArg.getWrappedObject();
         }
 
-        Map<String, Map<String, Map<Long, Map<String, TechUsageStatSum>>>> techStatsMap = getTechStatsMap();
+        Map<String, Map<String, Map<Long, Map<String, TechReportService.TechUsageStatSum>>>> techStatsMap = getTechStatsMap();
 
         ExecutionStatistics.get().end(NAME);
         return techStatsMap;
@@ -93,31 +87,31 @@ public class SortTechUsageStatsMethod implements WindupFreeMarkerMethod
     /**
      * Prepares a precomputed matrix - map of maps of maps: rowTag -> boxTag -> project -> silly label -> TechUsageStatSum.
      */
-    private Map<String, Map<String, Map<Long, Map<String, TechUsageStatSum>>>> getTechStatsMap()
+    private Map<String, Map<String, Map<Long, Map<String, TechReportService.TechUsageStatSum>>>> getTechStatsMap()
     {
-        Map<String, Map<String, Map<Long, Map<String, TechUsageStatSum>>>> map = new HashMap<>();
+        Map<String, Map<String, Map<Long, Map<String, TechReportService.TechUsageStatSum>>>> map = new HashMap<>();
 
         final Iterable<TechnologyUsageStatisticsModel> statModels = graphContext.service(TechnologyUsageStatisticsModel.class).findAll();
         for (TechnologyUsageStatisticsModel stat : statModels)
         {
             LOG.info(String.format("    Rolling up '%s', count: %sx, tags: %s", stat.getName(), stat.getOccurrenceCount(), stat.getTags()) );
 
-            final Set<String>[] normalAndSilly = splitSillyTagNames(graphContext, stat.getTags());
-            TechReportPlacement placement = processSillyLabels(graphContext, normalAndSilly[1]);
-            placement = normalizeSillyPlacement(graphContext, placement);
+            final Set<String>[] normalAndSilly = TechReportService.splitSillyTagNames(graphContext, stat.getTags());
+            TechReportService.TechReportPlacement placement = TechReportService.processSillyLabels(graphContext, normalAndSilly[1]);
+            placement = TechReportService.normalizeSillyPlacement(graphContext, placement);
 
             // Sort them out to the map.
-            final Map<String, Map<Long, Map<String, TechUsageStatSum>>> row = map.computeIfAbsent(placement.row.getName(), k -> new HashMap());
-            final Map<Long, Map<String, TechUsageStatSum>> box = row.computeIfAbsent(placement.box.getName(), k -> new HashMap<>());
+            final Map<String, Map<Long, Map<String, TechReportService.TechUsageStatSum>>> row = map.computeIfAbsent(placement.row.getName(), k -> new HashMap());
+            final Map<Long, Map<String, TechReportService.TechUsageStatSum>> box = row.computeIfAbsent(placement.box.getName(), k -> new HashMap<>());
 
             final String statsModelLabel = stat.getName();
             // All Projects
-            final Map<String, TechUsageStatSum> statSumAll = box.computeIfAbsent(Long.valueOf(0), k -> new HashMap<>());
-            statSumAll.put(statsModelLabel, new TechUsageStatSum(stat));
+            final Map<String, TechReportService.TechUsageStatSum> statSumAll = box.computeIfAbsent(Long.valueOf(0), k -> new HashMap<>());
+            statSumAll.put(statsModelLabel, new TechReportService.TechUsageStatSum(stat));
             // Respective project
             final Long projectKey = (Long) stat.getProjectModel().asVertex().getId();
-            final Map<String, TechUsageStatSum> statSum = box.computeIfAbsent(projectKey, k -> new HashMap<>());
-            statSum.put(statsModelLabel, new TechUsageStatSum(stat));
+            final Map<String, TechReportService.TechUsageStatSum> statSum = box.computeIfAbsent(projectKey, k -> new HashMap<>());
+            statSum.put(statsModelLabel, new TechReportService.TechUsageStatSum(stat));
         }
         return map;
     }
@@ -125,20 +119,20 @@ public class SortTechUsageStatsMethod implements WindupFreeMarkerMethod
     /**
      * A helper method to query the structure created above, since Freemarker can't query maps with numerical keys.
      */
-    private Map<String, TechUsageStatSum> queryMap(List arguments)
+    private Map<String, TechReportService.TechUsageStatSum> queryMap(List arguments)
     {
-        Map<String, Map<String, Map<Long, Map<String, TechUsageStatSum>>>> map = (Map<String, Map<String, Map<Long, Map<String, TechUsageStatSum>>>>) ((StringModel)arguments.get(0)).getWrappedObject();
+        Map<String, Map<String, Map<Long, Map<String, TechReportService.TechUsageStatSum>>>> map = (Map<String, Map<String, Map<Long, Map<String, TechReportService.TechUsageStatSum>>>>) ((StringModel)arguments.get(0)).getWrappedObject();
         String rowTagName = ((SimpleScalar) arguments.get(1)).getAsString();
         String boxTagName = ((SimpleScalar) arguments.get(2)).getAsString();
         Long projectId = ((SimpleNumber) arguments.get(3)).getAsNumber().longValue();
 
-        final Map<String, Map<Long, Map<String, TechUsageStatSum>>> rowMap = map.get(rowTagName);
+        final Map<String, Map<Long, Map<String, TechReportService.TechUsageStatSum>>> rowMap = map.get(rowTagName);
         if (null == rowMap)
             return null;
-        final Map<Long, Map<String, TechUsageStatSum>> boxMap = rowMap.get(boxTagName);
+        final Map<Long, Map<String, TechReportService.TechUsageStatSum>> boxMap = rowMap.get(boxTagName);
         if (null == boxMap)
             return null;
-        final Map<String, TechUsageStatSum> projectMap = boxMap.get(projectId);
+        final Map<String, TechReportService.TechUsageStatSum> projectMap = boxMap.get(projectId);
         return projectMap;
     }
